Uniformize DAQ_Move and DAQ_Viewer UI base classes
- Remove redundant ini_actuator_pb from DAQ_Move_UI_Base; use
  _setup_init_action + update_init_icon from ControlModuleUI
- Fix show_graph closeEvent bug (True → False) in move UI
- Rename show_tree → _show_settings in move UI for consistency
- Rename _tree → _settings_widget in move UI to match viewer
- Move enable_actions, _show_settings, show_settings to
  ControlModuleUI base so both UIs share the same implementation
- Add statusbar and _settings_widget None-defaults to base __init__
- Remove redundant command_sig redeclaration from DAQ_Viewer_UI
- Remove duplicate _ini_state = False assignments now covered by base
